Add --safe option to check all results with two hashes
I also swapped out the existing hash algorithm for a library called
xxhash, since I don't need a cryptographically-secure hash, and speed is
more important for this use case. The additional hash used in "safe"
mode is crc32. When concatenating hashes in a directory and hashing the
results, I opted to always use two algorithms, since the hash input is
not likely to be large.