Fix logic for simplifying results
I had previously broken logic for simplifying results
(i.e. not reporting child directories of duplicate directories)
during a refactor. This fixes that, and fixes another bug
that has always been present, where if B1 and B2 are duplicates,
but B1/C1, B2/C2, and C3 are duplicates, we would not indicate
that C3 was a duplicate of anything. Now we report B1/C1 and C3
being duplicates, but exclude B2/C2 from the results, so that my
existing logic for calculating potential space savings will continue
to work.